LPS-78225 Move methods used only by journal-web into journal-web
