+++ modules/apps/web-experience/journal/journal-web/build.gradle
@@ -14,6 +14,7 @@ dependencies {
 	provided project(":apps:collaboration:item-selector:item-selector-api")
 	provided project(":apps:collaboration:item-selector:item-selector-criteria-api")
 	provided project(":apps:collaboration:message-boards:message-boards-api")
+	provided project(":apps:collaboration:subscription:subscription-api")
 	provided project(":apps:collaboration:upload:upload-api")
 	provided project(":apps:forms-and-workflow:dynamic-data-mapping:dynamic-data-mapping-api")
 	provided project(":apps:forms-and-workflow:dynamic-data-mapping:dynamic-data-mapping-service")
