+++ modules/apps/web-experience/journal/journal-web/src/main/java/com/liferay/journal/web/internal/portlet/action/ActionUtil.java
@@ -27,9 +27,10 @@ import com.liferay.journal.service.JournalArticleServiceUtil;
 import com.liferay.journal.service.JournalFeedServiceUtil;
 import com.liferay.journal.service.JournalFolderServiceUtil;
 import com.liferay.journal.util.comparator.ArticleVersionComparator;
-import com.liferay.journal.util.impl.JournalUtil;
 import com.liferay.journal.web.internal.portlet.JournalPortlet;
 import com.liferay.journal.web.internal.security.permission.resource.JournalPermission;
+import com.liferay.journal.web.internal.util.JournalHelperUtil;
+import com.liferay.journal.web.util.JournalUtil;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.diff.CompareVersionsException;
 import com.liferay.portal.kernel.exception.PortalException;
@@ -140,7 +141,7 @@ public class ActionUtil {
 		String diffHtmlResults = null;
 
 		try {
-			diffHtmlResults = JournalUtil.diffHtml(
+			diffHtmlResults = JournalHelperUtil.diffHtml(
 				groupId, articleId, sourceVersion, targetVersion, languageId,
 				new PortletRequestModel(renderRequest, renderResponse),
 				themeDisplay);
